requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
torch==1.8.1
torchvision==0.9.1
opencv-python-headless==4.5.2.52
Pillow==8.2.0
numpy==1.20.3